Create README.md
This work is a b2b site work. The work was started by designing the backend first. After all the details were equipped with php, the frontend design was made. The frontend design was made with angular. Many features such as product sales, adding to cart, categories, in-page search, dealer entries, special discount coupons for dealers were added. In short, it meets most of the features expected from a b2b site.